+++ packages/happn-3/RELEASES.md
@@ -1057,6 +1057,11 @@
 -----------------
   - doc: added documentation regarding primus opts
 
-13.9.2 2023-02-18
+13.10.0 2023-03-03
+-----------------
+  - feat: token revocation and logout
+  - fix: token revocation cluster fix
+
+13.10.1 2023-03-04
 -----------------
   - Mongo-based cluster membership - dependencies update